Chore: store dir refactoring
