--- src/stores/types.ts
@@ -1,9 +1,0 @@
-export interface IPost {
-  seq: number
-  creator: string
-  title: string
-  body: string
-  imgList: string
-  viewCnt: number
-  startDt: string
-}
